Pylint
- Updated tests to adhere to pylint with docstrings and dividing tests to separate files
